bucketlist-api/requirements.txt
alembic==0.9.3
astroid==1.5.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.4.1
coveralls==1.1
cryptography==1.7.2
docopt==0.6.2
flasgger==0.6.6
Flask==0.12.2
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
idna==2.5
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
nose==1.3.7
passlib==1.7.1
pbr==3.1.1
psycopg2==2.7.1
pyasn1==0.3.1
pycparser==2.18
PyJWT==1.5.2
pylint==1.7.2
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.2
six==1.10.0
SQLAlchemy==1.1.11
stevedore==1.23.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
virtualenvwrapper-win==1.2.1
Werkzeug==0.12.2
wrapt==1.10.10
